+++ fractal_tasks_core/dev/lib_args_schemas.py
@@ -5,6 +5,7 @@ Copyright 2022 (C)
 
     Original authors:
     Tommaso Comparin <tommaso.comparin@exact-lab.it>
+    Yuri Chiucconi <yuri.chiucconi@exact-lab.it>
 
     This file is part of Fractal and was originally developed by eXact lab
     S.r.l.  <exact-lab.it> under contract with Liberali Lab from the Friedrich
@@ -13,25 +14,44 @@ Copyright 2022 (C)
 
 Helper functions to handle JSON schemas for task arguments.
 """
-import ast
 from pathlib import Path
 from typing import Any
+from typing import Optional
 
-from docstring_parser import parse
 from pydantic.decorator import ALT_V_ARGS
 from pydantic.decorator import ALT_V_KWARGS
 from pydantic.decorator import V_DUPLICATE_KWARGS
 from pydantic.decorator import V_POSITIONAL_ONLY_NAME
 from pydantic.decorator import ValidatedFunction
 
-import fractal_tasks_core
+from fractal_tasks_core.dev.lib_descriptions import (
+    _get_class_attrs_descriptions,
+)
+from fractal_tasks_core.dev.lib_descriptions import (
+    _get_function_args_descriptions,
+)
+from fractal_tasks_core.dev.lib_descriptions import (
+    _insert_class_attrs_descriptions,
+)
+from fractal_tasks_core.dev.lib_descriptions import (
+    _insert_function_args_descriptions,
+)
 from fractal_tasks_core.dev.lib_signature_constraints import _extract_function
 from fractal_tasks_core.dev.lib_signature_constraints import (
     _validate_function_signature,
 )
+from fractal_tasks_core.dev.lib_titles import _include_titles
 
 
 _Schema = dict[str, Any]
+
+FRACTAL_TASKS_CORE_PYDANTIC_MODELS = [
+    ("fractal_tasks_core", "lib_channels.py", "OmeroChannel"),
+    ("fractal_tasks_core", "lib_channels.py", "Window"),
+    ("fractal_tasks_core", "lib_input_models.py", "Channel"),
+    ("fractal_tasks_core", "lib_input_models.py", "NapariWorkflowsInput"),
+    ("fractal_tasks_core", "lib_input_models.py", "NapariWorkflowsOutput"),
+]
 
 
 def _remove_args_kwargs_properties(old_schema: _Schema) -> _Schema:
@@ -78,58 +98,25 @@ def _remove_pydantic_internals(old_schema: _Schema) -> _Schema:
     return new_schema
 
 
-def _get_args_descriptions(executable) -> dict[str, str]:
-    """
-    Extract argument descriptions for a task function
-    """
-    # Read docstring (via ast)
-    module_path = Path(fractal_tasks_core.__file__).parent / executable
-    module_name = module_path.with_suffix("").name
-    tree = ast.parse(module_path.read_text())
-    function = next(
-        f
-        for f in ast.walk(tree)
-        if (isinstance(f, ast.FunctionDef) and f.name == module_name)
-    )
-    docstring = ast.get_docstring(function)
-    # Parse docstring (via docstring_parser) and prepare output
-    parsed_docstring = parse(docstring)
-    descriptions = {
-        param.arg_name: param.description.replace("\n", " ")
-        for param in parsed_docstring.params
-    }
-    return descriptions
-
-
-def _include_args_descriptions_in_schema(*, schema, descriptions):
-    """
-    Merge the descriptions obtained via `_get_args_descriptions` into an
-    existing JSON Schema for task arguments
-    """
-    new_schema = schema.copy()
-    new_properties = schema["properties"].copy()
-    for key, value in schema["properties"].items():
-        if "description" in value:
-            raise ValueError("Property already has description")
-        else:
-            if key in descriptions:
-                value["description"] = descriptions[key]
-            else:
-                value["description"] = "Missing description"
-            new_properties[key] = value
-    new_schema["properties"] = new_properties
-    return new_schema
-
-
 def create_schema_for_single_task(
     executable: str,
-    package: str = "fractal_tasks_core.tasks",
+    package: str = "fractal_tasks_core",
+    custom_pydantic_models: Optional[list[tuple[str, str, str]]] = None,
 ) -> _Schema:
     """
     Main function to create a JSON Schema of task arguments
     """
+
+    # Extract the function name. Note: this could be made more general, but for
+    # the moment we assume the function has the same name as the module)
+    function_name = Path(executable).with_suffix("").name
+
     # Extract function from module
-    task_function = _extract_function(executable=executable, package=package)
+    task_function = _extract_function(
+        package_name=package,
+        module_relative_path=executable,
+        function_name=function_name,
+    )
 
     # Validate function signature against some custom constraints
     _validate_function_signature(task_function)
@@ -140,10 +127,48 @@ def create_schema_for_single_task(
     schema = _remove_args_kwargs_properties(schema)
     schema = _remove_pydantic_internals(schema)
 
-    # Include arg descriptions
-    descriptions = _get_args_descriptions(executable)
-    schema = _include_args_descriptions_in_schema(
-        schema=schema, descriptions=descriptions
+    # Include titles for custom-model-typed arguments
+    schema = _include_titles(schema)
+
+    # Include descriptions of function arguments
+    function_args_descriptions = _get_function_args_descriptions(
+        package_name=package,
+        module_relative_path=executable,
+        function_name=function_name,
     )
+    schema = _insert_function_args_descriptions(
+        schema=schema, descriptions=function_args_descriptions
+    )
+
+    # Merge lists of fractal-tasks-core and user-provided Pydantic models
+    user_provided_models = custom_pydantic_models or []
+    pydantic_models = FRACTAL_TASKS_CORE_PYDANTIC_MODELS + user_provided_models
+
+    # Check that model names are unique
+    tmp_class_names = set()
+    duplicate_class_names = set(
+        item[2]
+        for item in pydantic_models
+        if (item[2] in tmp_class_names or tmp_class_names.add(item[2]))
+    )
+    if duplicate_class_names:
+        pydantic_models_str = "  " + "\n  ".join(map(str, pydantic_models))
+        raise ValueError(
+            "Cannot parse docstrings for models with non-unique names "
+            f"{duplicate_class_names}, in\n{pydantic_models_str}"
+        )
+
+    # Extract model-attribute descriptions and insert them into schema
+    for package_name, module_relative_path, class_name in pydantic_models:
+        attrs_descriptions = _get_class_attrs_descriptions(
+            package_name=package_name,
+            module_relative_path=module_relative_path,
+            class_name=class_name,
+        )
+        schema = _insert_class_attrs_descriptions(
+            schema=schema,
+            class_name=class_name,
+            descriptions=attrs_descriptions,
+        )
 
     return schema