+++ fractal_tasks_core/lib_input_models.py
@@ -26,12 +26,15 @@ class Channel(BaseModel):
     """
 
     wavelength_id: Optional[str] = None
+    """Unique ID for the channel wavelength, e.g. ``A01_C01``."""
+
     label: Optional[str] = None
+    """Name of the channel"""
 
     @validator("label", always=True)
     def mutually_exclusive_channel_attributes(cls, v, values):
         """
-        Attributes ``label`` and ``wavelength_id`` are mutually exclusive.
+        Check that either ``label`` or ``wavelength_id`` is set.
         """
         wavelength_id = values.get("wavelength_id")
         label = v
@@ -53,11 +56,19 @@ class NapariWorkflowsInput(BaseModel):
     """
 
     type: Literal["image", "label"]
+    """Input type (supported: ``image`` or ``label``)."""
+
     label_name: Optional[str]
+    """Label name (for label inputs only)."""
+
     channel: Optional[Channel]
+    """Channel object (for image inputs only)."""
 
     @validator("label_name", always=True)
     def label_name_is_present(cls, v, values):
+        """
+        Check that label inputs have ``label_name`` set.
+        """
         _type = values.get("type")
         if _type == "label" and not v:
             raise ValueError(
@@ -67,6 +78,9 @@ class NapariWorkflowsInput(BaseModel):
 
     @validator("channel", always=True)
     def channel_is_present(cls, v, values):
+        """
+        Check that image inputs have ``channel`` set.
+        """
         _type = values.get("type")
         if _type == "image" and not v:
             raise ValueError(f"Input item has type={_type} but channel={v}.")
@@ -79,11 +93,19 @@ class NapariWorkflowsOutput(BaseModel):
     """
 
     type: Literal["label", "dataframe"]
+    """Output type (supported: ``label`` or ``dataframe``)."""
+
     label_name: Optional[str] = None
+    """Label name (for label outputs only)."""
+
     table_name: Optional[str] = None
+    """Table name (for dataframe outputs only)."""
 
     @validator("label_name", always=True)
     def label_name_only_for_label_type(cls, v, values):
+        """
+        Check that label_name is set only for label outputs.
+        """
         _type = values.get("type")
         if (_type == "label" and (not v)) or (_type != "label" and v):
             raise ValueError(
@@ -93,6 +115,9 @@ class NapariWorkflowsOutput(BaseModel):
 
     @validator("table_name", always=True)
     def table_name_only_for_dataframe_type(cls, v, values):
+        """
+        Check that table_name is set only for dataframe outputs.
+        """
         _type = values.get("type")
         if (_type == "dataframe" and (not v)) or (_type != "dataframe" and v):
             raise ValueError(