+++ tests/tasks/test_valid_args_schemas.py
@@ -99,7 +99,8 @@ def test_task_functions_have_valid_signatures():
     Test that task functions have valid signatures.
     """
     for ind_task, task in enumerate(TASK_LIST):
-        task_function = _extract_function(task["executable"])
+        function_name = Path(task["executable"]).with_suffix("").name
+        task_function = _extract_function(task["executable"], function_name)
         _validate_function_signature(task_function)
 
 
@@ -111,6 +112,10 @@ def test_args_schemas_are_up_to_date():
         print(f"Now handling {task['executable']}")
         old_schema = TASK_LIST[ind_task]["args_schema"]
         new_schema = create_schema_for_single_task(task["executable"])
+        # The following step is required because some arguments may have a
+        # default which has a non-JSON type (e.g. a tuple), which we need to
+        # convert to JSON type (i.e. an array) before comparison.
+        new_schema = json.loads(json.dumps(new_schema))
         assert new_schema == old_schema
 
 
@@ -130,3 +135,43 @@ def test_args_schema_comply_with_jsonschema_specs(jsonschema_validator):
             f"Schema for task {task['executable']} is valid for "
             f"{jsonschema_validator}."
         )
+
+
+def test_args_title():
+    """
+    Check that two kinds of properties have the correct title set:
+    1. Task arguments which have a custom-model type
+    2. Custom-model-typed attributes of custom-model-typed task arguments
+
+    See
+    https://github.com/fractal-analytics-platform/fractal-tasks-core/issues/446
+    """
+
+    cellpose_task = next(
+        task for task in TASK_LIST if task["name"] == "Cellpose Segmentation"
+    )
+    new_schema = create_schema_for_single_task(cellpose_task["executable"])
+    properties = new_schema["properties"]
+    # Standard task argument
+    level_prop = properties["level"]
+    debug(level_prop)
+    assert level_prop["title"] == "Level"
+    # Custom-model-typed task argument
+    channel2_prop = properties["channel2"]
+    debug(channel2_prop)
+    assert channel2_prop["title"] == "Channel2"
+
+    create_ome_zarr_task = next(
+        task
+        for task in TASK_LIST
+        if task["name"] == "Create OME-Zarr structure"
+    )
+    new_schema = create_schema_for_single_task(
+        create_ome_zarr_task["executable"]
+    )
+    definitions = new_schema["definitions"]
+    omero_channel_def = definitions["OmeroChannel"]
+    # Custom-model-typed attribute of custom-model-typed task argument
+    window_prop = omero_channel_def["properties"]["window"]
+    debug(window_prop)
+    assert window_prop["title"] == "Window"